80 service tests passed out of 99

+++ src/main/java/za/ac/cput/service/ReviewService.java
@@ -3,6 +3,7 @@ package za.ac.cput.service;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import za.ac.cput.domain.Review;
 import za.ac.cput.repository.ReviewRepository;
 
@@ -16,6 +17,7 @@ import java.util.List;
  * @date 24-Sep-24
  */
 @Slf4j
+@Transactional
 @Service
 public class ReviewService implements IReview {
 
@@ -28,16 +30,19 @@ public class ReviewService implements IReview {
     }
 
     @Override
+    @Transactional(readOnly = false)
     public Review create(Review review) {
         return repository.save(review);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public Review read(Long id) {
         return repository.findById(id).orElse(null);
     }
 
     @Override
+    @Transactional(readOnly = false)
     public Review update(Review review) {
         Review existingReview = repository.findById(review.getId()).orElse(null);
         if (existingReview != null) {
@@ -58,33 +63,36 @@ public class ReviewService implements IReview {
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Review> findAll() {
         return repository.findAll();
     }
 
     @Override
+    @Transactional(readOnly = false)
     public boolean delete(Long id) {
-        repository.deleteById(id);
-
-        // Check if the entity still exists after deletion
-        boolean exists = repository.existsById(id);
-
-        // Return false if entity was deleted successfully, otherwise return true
-        return !exists;
+        if (repository.existsById(id)) {
+            repository.deleteById(id);
+            return true;
+        }
+        log.warn("Attempt to delete a non-existent review with ID: ", id);
+        return false;
     }
 
-
     @Override
+    @Transactional(readOnly = true)
     public List<Review> findByProduct_Id(Long id) {
         return repository.findByProduct_Id(id);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Review> findByRating(int rating) {
         return repository.findByRating(rating);
     }
 
     @Override
+    @Transactional(readOnly = true)
     public List<Review> findByRatingGreaterThan(int rating) {
         return repository.findByRatingGreaterThan(rating);
     }
